+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2018-12-06
+### Added
+
 ## [0.3.0] - 2018-12-05
 ### Changed
 - `SQLiteCursor#getBlob()` now returns a `Uint8List`, not a `ByteBuffer`
@@ -112,6 +115,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SQLiteDatabase#rawQuery()` method
 - `SQLiteDatabase#setLocale()` method
 
+[0.3.1]:  https://github.com/drydart/flutter_sqlcipher/compare/0.3.0...0.3.1
 [0.3.0]:  https://github.com/drydart/flutter_sqlcipher/compare/0.2.5...0.3.0
 [0.2.5]:  https://github.com/drydart/flutter_sqlcipher/compare/0.2.4...0.2.5
 [0.2.4]:  https://github.com/drydart/flutter_sqlcipher/compare/0.2.3...0.2.4